+++ include/slang/util/Hash.h
@@ -16,7 +16,7 @@ namespace slang {
 
 /// Hashes the provided input using the xxhash XXH3 algorithm.
 inline size_t xxhash(const void* input, size_t len) {
-    return XXH3_64bits(input, len);
+    return (size_t)XXH3_64bits(input, len);
 }
 
 inline void hash_combine(size_t&) {
